ConversationInterface file
