+++ Wire-iOS/Sources/Protocols/ZMUserSessionInterface.swift
