+++ Wire-iOS/Sources/UserInterface/ConversationList/Container/ViewModel/ConversationListViewControllerViewModel+ConversationListContentDelegate.swift
@@ -47,21 +47,3 @@ extension ConversationListViewController.ViewModel {
         actionsController?.presentMenu(from: view, context: .list)
     }
 }
-
-protocol ConversationInterface: class {
-    var conversationType: ZMConversationType { get }
-    var teamRemoteIdentifier: UUID? { get set }
-    var connectedUser: ZMUser? { get }
-    var displayName: String { get }
-    var isArchived: Bool { get set }
-    var isReadOnly: Bool { get }
-    var isFavorite: Bool { get }
-    var mutedMessageTypes: MutedMessageTypes { get set }
-    var activeParticipants: Set<ZMUser> { get }
-    var folderName: String? { get }
-    var unreadMessages: [ZMConversationMessage] { get }
-
-    func canMarkAsUnread() -> Bool
-}
-
-extension ZMConversation: ConversationInterface {}
